Feature: normalize mapping data in entity
