+++ src/Mapping/Validator/NullValidator.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Apitte\Core\Mapping\Validator;
+
+class NullValidator implements IEntityValidator
+{
+
+	/**
+	 * @param object $entity
+	 * @return void
+	 */
+	public function validate($entity)
+	{
+		// Hell nothing..
+	}
+
+}
